Separate inputs and cycle methods for units and layers
Before:
>>> g_e = 0.5
>>> unit.cycle(g_e)

After:
>>> g_e = 0.5
>>> unit.add_excitatory(g_e)
>>> unit.cycle()

Same for layers, with array of inputs.

This allows connections to add inputs to a unit without the need
for the Network to have to gather excitatory inputs and compute
netraw, and reshape layer inputs. Not sure if this will still
work if we need to implement complex scaling methods.